begin guttannen calibration

+++ src/utils/LOO.py
@@ -115,7 +115,7 @@ if __name__ == "__main__":
     results_list = manager.list()
 
     # Create process pool with four processes
-    num_processes = multiprocessing.cpu_count()/2
+    num_processes = int(multiprocessing.cpu_count()/2)
     pool = multiprocessing.Pool(processes=num_processes)
     processes = []
